+++ ethereum/jsonrpc/src/main/java/tech/pegasys/pantheon/ethereum/jsonrpc/internal/methods/NetVersion.java
@@ -16,6 +16,8 @@ import tech.pegasys.pantheon.ethereum.jsonrpc.internal.JsonRpcRequest;
 import tech.pegasys.pantheon.ethereum.jsonrpc.internal.response.JsonRpcResponse;
 import tech.pegasys.pantheon.ethereum.jsonrpc.internal.response.JsonRpcSuccessResponse;
 
+import java.math.BigInteger;
+
 /**
  * In Consensys' client, net_version maps to the network id, as specified in *
  * https://github.com/ethereum/wiki/wiki/JSON-RPC#net_version
@@ -25,7 +27,7 @@ import tech.pegasys.pantheon.ethereum.jsonrpc.internal.response.JsonRpcSuccessRe
 public class NetVersion implements JsonRpcMethod {
   private final String chainId;
 
-  public NetVersion(final int chainId) {
+  public NetVersion(final BigInteger chainId) {
     this.chainId = String.valueOf(chainId);
   }
